Add wrapper <div> to fix zoom issue on mobile
Since the #grid <div> is very wide, mobile browsers make the root scroller very tall so that when the user zooms out to fit the entire #grid into the visible viewport, the root scroller still covers the entire height of the visual viewport. The fix is to horizontally scroll the #grid inside a wrapper <div> instead.