fix: update switch based on fullscreenchange event
The bug: When the user enters full screen and then clicks a link to Deku Deals, the browser will exit full screen and navigate to that page. When the uesr then navigates back, the switch will still be checked even though the page is not in full screen.

The fix updates the switch on every fullscreenchange event.