show actual  prices
Animal Crossing S2 edition is 65 and its upgrade pack 5, so I’m switching to showing the actual prices (there is no Unicode character for 1/2 in a circle).

Also, instead of NSO I show an asterisk now. It’s simpler. E.g. 0* means the game is free but requires an NSO subscription.